[AWIBOF-7396] Change prefix mf_ to meta_file_
