+++ src/meta_file_intf/meta_file_intf.h
@@ -37,7 +37,7 @@
 #include "src/meta_file_intf/async_context.h"
 #include "src/meta_file_intf/meta_file_include.h"
 #include "src/metafs/include/meta_volume_type.h"
-#include "src/metafs/include/mf_property.h"
+#include "src/metafs/include/meta_file_property.h"
 
 namespace pos
 {
