--- src/metafs/include/mf_property.h
@@ -1,71 +1,0 @@
-/*
- *   BSD LICENSE
- *   Copyright (c) 2021 Samsung Electronics Corporation
- *   All rights reserved.
- *
- *   Redistribution and use in source and binary forms, with or without
- *   modification, are permitted provided that the following conditions
- *   are met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in
- *       the documentation and/or other materials provided with the
- *       distribution.
- *     * Neither the name of Samsung Electronics Corporation nor the names of
- *       its contributors may be used to endorse or promote products derived
- *       from this software without specific prior written permission.
- *
- *   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- *   "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- *   LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- *   A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- *   OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- *   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- *   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- *   DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- *   THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- *   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- *   OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#pragma once
-
-#include "mf_integrity_type.h"
-
-namespace pos
-{
-// definitions below indicate io characteristics of the corresponding meta file access
-// meta filesystem stores metadata and service io efficiently according to the setting given by user
-enum class MetaFileType
-{
-    SpecialPurposeMap,
-    Journal,
-    Map,
-    General,
-    MAX
-};
-
-class MetaFilePropertySet
-{
-public:
-    MetaFilePropertySet(void)
-    : MetaFilePropertySet(MetaFileIntegrityType::Default, MetaFileType::General)
-    {
-    }
-    MetaFilePropertySet(MetaFileType type)
-    : MetaFilePropertySet(MetaFileIntegrityType::Default, type)
-    {
-    }
-    MetaFilePropertySet(MetaFileIntegrityType integrity, MetaFileType type)
-    : integrity(integrity),
-      type(type)
-    {
-    }
-    ~MetaFilePropertySet(void) = default;
-
-    MetaFileIntegrityType integrity;
-    MetaFileType type;
-};
-} // namespace pos
