+++ src/metafs/mai/metafs_wbt_api.h
@@ -35,7 +35,7 @@
 #include <string>
 #include <vector>
 
-#include "src/metafs/include/mf_dataformat.h"
+#include "src/metafs/include/meta_file_dataformat.h"
 #include "src/metafs/mai/metafs_file_control_api.h"
 #include "src/metafs/mai/metafs_management_api.h"
 #include "mk/ibof_config.h"
