+++ src/metafs/mvm/volume/inode_deleter.h
@@ -35,7 +35,6 @@
 #include "src/metafs/include/metafs_control_request.h"
 #include "src/include/pos_event_id.h"
 #include "src/metafs/common/metafs_type.h"
-#include "src/metafs/mvm/volume/mf_inode.h"
 
 #include <utility>
