+++ src/wbt/wbt_metafs_cmd_handler.h
@@ -44,7 +44,7 @@
 #include "nlohmann/json.hpp"
 #include "src/helper/json/json_helper.h"
 #include "src/metafs/metafs.h"
-#include "src/metafs/include/mf_dataformat.h"
+#include "src/metafs/include/meta_file_dataformat.h"
 
 namespace pos
 {
