+++ test/unit-tests/metafs/include/CMakeLists.txt
@@ -1,13 +1,12 @@
 POS_ADD_UNIT_TEST(os_header_ut os_header_test.cpp)
 POS_ADD_UNIT_TEST(meta_storage_specific_ut meta_storage_specific_test.cpp)
-POS_ADD_UNIT_TEST(mf_property_ut mf_property_test.cpp)
+POS_ADD_UNIT_TEST(meta_file_property_ut meta_file_property_test.cpp)
 POS_ADD_UNIT_TEST(meta_storage_info_ut meta_storage_info_test.cpp)
 POS_ADD_UNIT_TEST(meta_volume_type_ut meta_volume_type_test.cpp)
-POS_ADD_UNIT_TEST(mf_dataformat_ut mf_dataformat_test.cpp)
-POS_ADD_UNIT_TEST(mf_lock_type_ut mf_lock_type_test.cpp)
+POS_ADD_UNIT_TEST(meta_file_dataformat_ut meta_file_dataformat_test.cpp)
 POS_ADD_UNIT_TEST(metafs_aiocb_cxt_ut metafs_aiocb_cxt_test.cpp)
 POS_ADD_UNIT_TEST(metafs_common_const_ut metafs_common_const_test.cpp)
 POS_ADD_UNIT_TEST(metafs_file_control_request_ut metafs_file_control_request_test.cpp)
-POS_ADD_UNIT_TEST(mf_integrity_type_ut mf_integrity_type_test.cpp)
+POS_ADD_UNIT_TEST(meta_file_integrity_type_ut meta_file_integrity_type_test.cpp)
 POS_ADD_UNIT_TEST(meta_file_extent_ut meta_file_extent_test.cpp)
 POS_ADD_UNIT_TEST(metafs_service_ut metafs_service_test.cpp)
