+++ test/unit-tests/metafs/include/meta_file_dataformat_mock.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "src/metafs/include/mf_dataformat.h"
+#include "src/metafs/include/meta_file_dataformat.h"
 
 namespace pos
 {
