+++ test/unit-tests/metafs/include/meta_file_integrity_type_test.cpp
@@ -1,0 +1,3 @@
+#include "src/metafs/include/meta_file_integrity_type.h"
+
+#include <gtest/gtest.h>
