--- test/unit-tests/metafs/include/mf_integrity_type_mock.h
@@ -1,7 +1,0 @@
-#include <gmock/gmock.h>
-
-#include <list>
-#include <string>
-#include <vector>
-
-#include "src/metafs/include/mf_integrity_type.h"
