--- test/unit-tests/metafs/include/mf_integrity_type_test.cpp
@@ -1,3 +1,0 @@
-#include "src/metafs/include/mf_integrity_type.h"
-
-#include <gtest/gtest.h>
