--- test/unit-tests/metafs/include/mf_lock_type_test.cpp
@@ -1,3 +1,0 @@
-#include "src/metafs/include/mf_lock_type.h"
-
-#include <gtest/gtest.h>
