+++ test/unit-tests/metafs/mvm/volume/CMakeLists.txt
@@ -10,7 +10,7 @@ POS_ADD_UNIT_TEST(on_volume_meta_region_ut on_volume_meta_region_test.cpp)
 POS_ADD_UNIT_TEST(inode_table_ut inode_table_test.cpp)
 POS_ADD_UNIT_TEST(catalog_manager_ut catalog_manager_test.cpp)
 POS_ADD_UNIT_TEST(extent_allocator_ut extent_allocator_test.cpp)
-POS_ADD_UNIT_TEST(mf_inode_req_ut mf_inode_req_test.cpp)
+POS_ADD_UNIT_TEST(meta_file_inode_req_ut meta_file_inode_req_test.cpp)
 POS_ADD_UNIT_TEST(on_volume_meta_region_mgr_ut on_volume_meta_region_mgr_test.cpp)
-POS_ADD_UNIT_TEST(mf_inode_ut mf_inode_test.cpp)
+POS_ADD_UNIT_TEST(meta_file_inode_ut meta_file_inode_test.cpp)
 POS_ADD_UNIT_TEST(file_descriptor_allocator_ut file_descriptor_allocator_test.cpp)
