+++ test/unit-tests/metafs/mvm/volume/meta_file_inode_req_mock.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include "src/metafs/mvm/volume/mf_inode_req.h"
+#include "src/metafs/mvm/volume/meta_file_inode_req.h"
 
 namespace pos
 {
