+++ test/unit-tests/metafs/mvm/volume/meta_file_inode_req_test.cpp
@@ -1,4 +1,4 @@
-#include "src/metafs/mvm/volume/mf_inode_req.h"
+#include "src/metafs/mvm/volume/meta_file_inode_req.h"
 
 #include <gtest/gtest.h>
